+++ src/components/InputPassword/InputPassword.styles.ts
@@ -1,0 +1,10 @@
+import { css } from '@emotion/css';
+
+export const styles = {
+  innerButton: css`
+    padding: 0;
+  `,
+  icon: css`
+    display: block;
+  `,
+};