+++ src/components/Tooltip/index.js
@@ -103,7 +103,6 @@ export const withTooltip = (
 
     getWrapperElement(): ?HTMLElement {
       const { wrapperRef } = this;
-
       const componentRef = wrapperRef.current && ((wrapperRef.current: any): ComponentRef).elementRef;
       const domNodeRef = ((wrapperRef: any): { current: HTMLElement });
 
@@ -115,7 +114,6 @@ export const withTooltip = (
       const { tooltipRef } = this;
       const tooltipElement = tooltipRef.current;
       const wrapperElement = this.getWrapperElement();
-
       if (((visible && !prevState.visible) || prevProps !== this.props) && tooltipElement && wrapperElement) {
         const bodyWidth = document.body ? document.body.clientWidth : 0;
         const wrapperRect = wrapperElement.getBoundingClientRect();