+++ src/index.ts
@@ -8,5 +8,9 @@ export * from './components/UploadZone';
 export * from './components/SVGImage';
 export * from './components/Icon';
 export * from './components/Checkbox';
+export * from './components/InputPassword';
+export * from './components/Input';
+export * from './components/LabeledInput';
+export * from './components/ControlsPanel';
 
 export * from './variables';