+++ CHANGELOG.md
@@ -6,11 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+-- Refactor in TypeScript: Input, InputPassword, ControlPanel, LabeledInput;
 ## [0.50.11] - 2022-02-18
 
 - Added the cancelText field in ConfirmModal;
 
 - Improve the `Table` component;
+
 
 ## [0.50.10] - 2022-02-14